+++ toad/nn/trainer/__init__.py
@@ -1,10 +1,11 @@
-from .history import History
+from .history import History, get_current_history
 from .callback import callback
 from .earlystop import earlystopping
 from .trainer import Trainer
 
 __all__ = [
     'History',
+    'get_current_history',
     'callback',
     'earlystopping',
     'Trainer',