+++ toad/nn/trainer/history_test.py
@@ -1,6 +1,6 @@
 import torch
 import numpy as np
-from .history import History
+from .history import History, get_current_history
 
 def test_history_log():
     history = History()
@@ -11,3 +11,11 @@ def test_history_log():
     assert history['tensor'].shape == (30, 5)
 
 
+def test_current_history():
+    history = History()
+
+    with history:
+        h = get_current_history()
+        h.log('tensor', torch.rand(3, 5))
+    
+    assert history['tensor'].shape == (3, 5)