+++ toad/nn/trainer/trainer_test.py
@@ -1,9 +1,11 @@
-from toad.nn.trainer.history import History
+import pytest
+
 import torch
 from torch import nn
 import torch.nn.functional as F
 from torch.utils.data import TensorDataset, DataLoader
 
+from .history import History
 from ..module import Module
 from .trainer import Trainer
 from .callback import callback
@@ -58,6 +60,23 @@ def test_trainer_early_stopping():
     assert len(trainer.history) == 4
 
 
+def test_trainer_fit_step():
+    model = TestModel(NUM_FEATS, NUM_CLASSES)
+    trainer = Trainer(model, loader)
+    step_count = 0
+
+    @trainer.fit_step
+    def step(model, batch):
+        x, y = batch
+        y_hat = model(x)
+        nonlocal step_count
+        step_count += 1
+        return F.cross_entropy(y_hat, y)
+    
+    trainer.train(epoch = 2)
+    assert step_count > 1
+
+
 def test_multi_callbacks():
     log = {}
     
@@ -77,6 +96,22 @@ def test_multi_callbacks():
     assert len(log['loss']) == 157
 
 
+def test_trainer_evaluate():
+    model = TestModel(NUM_FEATS, NUM_CLASSES)
+    trainer = Trainer(model, loader)
+
+    @trainer.fit_step
+    def step(model, batch):
+        x, y = batch
+        y_hat = model(x)
+        return F.cross_entropy(y_hat, y)
+    
+    history = trainer.evaluate(loader)
+
+    assert len(history["loss"]) == 157
+    
+
+
 class TestModel2(TestModel):
     def fit_step(self, batch, loss=None):
         x, y = batch
@@ -89,3 +124,71 @@ def test_trainer_loss():
     trainer = Trainer(model, loader, loss = F.cross_entropy)
     trainer.train(epoch = 2)
     assert len(trainer.history) == 2
+
+
+# def test_trainer_distributed():
+#     model = TestModel(NUM_FEATS, NUM_CLASSES)
+#     trainer = Trainer(model, loader)
+#     trainer.distributed(workers = 2)
+#     trainer.train(epoch = 5)
+
+
+
+### distribut model test
+# from toad.nn.trainer.trainer import Trainer
+# from torchvision.transforms import ToTensor
+# import torch
+# from torch import nn
+# from torchvision import datasets
+# from toad.nn import Module
+# from torch.utils.data import DataLoader
+# import ray
+
+# class NeuralNetwork(Module):
+#     def __init__(self):
+#         super(NeuralNetwork, self).__init__()
+#         self.flatten = nn.Flatten()
+#         self.linear_relu_stack = nn.Sequential(
+#             nn.Linear(28 * 28, 512),
+#             nn.ReLU(),
+#             nn.Linear(512, 512),
+#             nn.ReLU(),
+#             nn.Linear(512, 10),
+#             nn.ReLU(),
+#         )
+#     def forward(self, x):
+#         x = self.flatten(x)
+#         logits = self.linear_relu_stack(x)
+#         return logits
+#     def fit_step(self, batch):
+#         X, y = batch
+#         pred =self(X)
+#         loss_fn=nn.CrossEntropyLoss()
+#         return loss_fn(pred, y)
+
+
+# @pytest.mark.skip("distributed trainer skip")
+# def test_distribute_example():
+#     training_data = datasets.FashionMNIST(
+#         root="~/data",
+#         train=True,
+#         download=True,
+#         transform=ToTensor(),
+#     )
+#     # Download test data from open datasets.
+#     test_data = datasets.FashionMNIST(
+#         root="~/data",
+#         train=False,
+#         download=True,
+#         transform=ToTensor(),
+#     )
+#     worker_batch_size = 64 // 4
+#     # Create data loaders.
+#     train_dataloader = DataLoader(training_data, batch_size=16)
+#     test_dataloader = DataLoader(test_data, batch_size=16)
+#     model=NeuralNetwork()
+#     optimizer = torch.optim.SGD(model.parameters(), lr=1e-3)
+#     trainer=Trainer(model,train_dataloader,optimizer)
+#     trainer.distributed(address="ray://172.20.144.21:10001",num_works=4,use_gpu=False)
+#     trainer.train(epoch=1)
+#     trainer.evaluate(test_dataloader)